Second version
It allows to save the result in .txt file, not only print the result on a console. This version includes info about the author of each message. In spite of the original Zoom format, the message author's name is not repeated if the author writes a few messages one by one. The code was reformatted, a few loops are changed from classic for-style to foreach-style. Minor bugs were removed as well.